Line엔티티 sections를 일급컬렉션으로 변경
